requirements.txt
streamlit 
tensorflow
numpy 
pillow